Feature: have been created: methods for serialization/deserialization, tests serialization/deserialization, crashed singleton and test for this, set random value for field with annotation and test for this, immutable class and test for this
