--- src/main/java/com/itea/homeworks/annotation/Information.java
@@ -1,13 +1,0 @@
-package com.itea.homeworks.annotation;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target(ElementType.FIELD)
-@Retention(RetentionPolicy.RUNTIME)
-public @interface Information {
-    int first() default 1;
-    int second() default 2;
-}
