--- src/main/java/com/itea/homeworks/collections/arraylist/ArrayListG.java
@@ -1,9 +1,0 @@
-package com.itea.homeworks.collections.arraylist;
-
-public interface ArrayListG<T> {
-    void add(T item);
-    void remove(int index);
-    Object size();
-    void clear();
-    Object get(int index);
-}
