--- src/main/java/com/itea/homeworks/collections/hashmap/HashMapG.java
@@ -1,13 +1,0 @@
-package com.itea.homeworks.collections.hashmap;
-
-public interface HashMapG<K,V> {
-    Object get(K key);
-
-    boolean put(K key, V value);
-
-    void clear();
-
-    void remove(K key);
-
-    int size();
-}
