--- src/main/java/com/itea/homeworks/exception/MyException.java
@@ -1,11 +1,0 @@
-package com.itea.homeworks.exception;
-
-public class MyException extends RuntimeException {
-    public MyException(){
-        super();
-    }
-
-    public MyException(String message) {
-        super(message);
-    }
-}
