--- src/main/java/com/itea/homeworks/singleton/Singleton.java
@@ -1,15 +1,0 @@
-package com.itea.homeworks.singleton;
-
-public class Singleton {
-    private Singleton instance;
-
-    private Singleton() {
-    }
-
-    public Singleton getInstance() {
-        if (instance == null){
-            return new Singleton();
-        }
-        return instance;
-    }
-}
