+++ src/main/java/com/itea/homeworks/task5/annotation/Information.java
@@ -1,0 +1,12 @@
+package com.itea.homeworks.task5.annotation;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target(ElementType.FIELD)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface Information {
+    int first();
+}
