+++ src/main/java/com/itea/homeworks/task5/collections/arraylist/ArrayListG.java
@@ -1,0 +1,9 @@
+package com.itea.homeworks.task5.collections.arraylist;
+
+public interface ArrayListG<T> {
+    void add(T item);
+    void remove(int index);
+    Object size();
+    void clear();
+    Object get(int index);
+}
