+++ src/main/java/com/itea/homeworks/task5/collections/hashmap/HashMapG.java
@@ -1,0 +1,13 @@
+package com.itea.homeworks.task5.collections.hashmap;
+
+public interface HashMapG<K,V> {
+    Object get(K key);
+
+    boolean put(K key, V value);
+
+    void clear();
+
+    void remove(K key);
+
+    int size();
+}
