+++ src/main/java/com/itea/homeworks/task5/exception/MyException.java
@@ -1,0 +1,11 @@
+package com.itea.homeworks.task5.exception;
+
+public class MyException extends RuntimeException {
+    public MyException(){
+        super();
+    }
+
+    public MyException(String message) {
+        super(message);
+    }
+}
