+++ src/main/java/com/itea/homeworks/task5/utils/calculation/OperType.java
@@ -1,0 +1,5 @@
+package com.itea.homeworks.task5.utils.calculation;
+
+public interface OperType {
+    double calc(double a, double b);
+}
