+++ src/main/java/com/itea/homeworks/task6/changefield/Children.java
@@ -1,0 +1,8 @@
+package com.itea.homeworks.task6.changefield;
+
+import com.itea.homeworks.task5.annotation.Information;
+
+public class Children {
+    @Information(first = 10)
+    public int ageFirstChildren;
+}
