+++ src/main/java/com/itea/homeworks/task6/serdeser/SerDeser.java
@@ -1,0 +1,25 @@
+package com.itea.homeworks.task6.serdeser;
+
+import java.io.FileInputStream;
+import java.io.FileOutputStream;
+import java.io.IOException;
+import java.io.ObjectOutputStream;
+import java.io.ObjectInputStream;
+
+public class SerDeser {
+    public void serialization(Object object, String path) throws IOException {
+        FileOutputStream fos = new FileOutputStream(path);
+        ObjectOutputStream oos = new ObjectOutputStream(fos);
+
+        oos.writeObject(object);
+        oos.flush();
+        oos.close();
+    }
+
+    public Object deserialization(String path) throws IOException, ClassNotFoundException {
+        FileInputStream fis = new FileInputStream(path);
+        ObjectInputStream ois = new ObjectInputStream(fis);
+
+        return ois.readObject();
+    }
+}
