--- src/main/java/com/itea/homeworks/utils/calculation/OperType.java
@@ -1,5 +1,0 @@
-package com.itea.homeworks.utils.calculation;
-
-public interface OperType {
-    double calc(double a, double b);
-}
